[lexical-playground] Refactor: Migrate TwitterPlugin to TwitterExtension
Converts the legacy React TwitterPlugin into a declarative
TwitterExtension using defineExtension, registering the TweetNode and
INSERT_TWEET_COMMAND through the extension system.
